+++ common/src/config/GameConfig.cpp
@@ -186,6 +186,7 @@ bool GameConfig::visit_vars(T&& visitor, bool is_save)
     result &= visitor(dash_faction_key, "Player Join Beep", player_join_beep);
     result &= visitor(dash_faction_key, "Autosave", autosave);
     result &= visitor(dash_faction_key, "Disguise as Alpine Faction", disguise_as_af);
+    result &= visitor(dash_faction_key, "Ignore Level Version", ignore_level_version);
 
     return result;
 }