+++ launcher/resource.h
@@ -74,6 +74,7 @@
 #define IDC_RENDERER_COMBO              1058
 #define IDC_DISGUISE_AS_AF_CHECK        1059
 #define IDC_MUZZLE_FLASH_CHECK          1060
+#define IDC_IGNORE_LEVEL_VERSION_CHECK  1061
 
 // Next default values for new objects
 //
@@ -81,7 +82,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        112
 #define _APS_NEXT_COMMAND_VALUE         32771
-#define _APS_NEXT_CONTROL_VALUE         1061
+#define _APS_NEXT_CONTROL_VALUE         1062
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif